fixing decaying tree code...

+++ twin_tree.c
@@ -29,7 +29,7 @@ int main(int argc, char* argv[])
     curr = stream;
     end = stream;
 
-    int* arr = (int*) malloc(DICT_SIZE * sizeof(int));
+    double* arr = (double*) malloc(DICT_SIZE * sizeof(double));
 
     data sorted = create_sorted_dummy();
